# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "共域 CoField API",
    "description": "以共同事件为核心的校园社交连接器。\n\n领域词汇见 CONTEXT.md；用户可见文案不使用这些词汇，映射见 docs/07。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/envelopes/{envelope_id}:revoke": {
      "post": {
        "tags": [
          "envelopes"
        ],
        "summary": "Revoke Envelope",
        "description": "收回授权。撤销后新请求立刻被拒，不用等任何异步清理。",
        "operationId": "revoke_envelope_api_envelopes__envelope_id__revoke_post",
        "parameters": [
          {
            "name": "envelope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Envelope Id"
            }
          },
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          },
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/events": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "My Events",
        "description": "我参加过的事，进行中、已完成、已取消都在。\n\n空列表不是错误：一个刚注册的人本来就什么都没参加过，\n这一屏该说的是\"先说说想做点什么\"，不是 404。",
        "operationId": "my_events_api_me_events_get",
        "parameters": [
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          },
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyEventOut"
                  },
                  "title": "Response My Events Api Me Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}/again": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Again",
        "description": "照这件事再来一次，先给一张草稿。",
        "operationId": "again_api_events__event_id__again_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          },
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgainOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}:again": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Ask Them Again",
        "description": "把这条新需求直接问上次那几个人。\n\n**不是把他们拉进来。** 不变量 3：成局前只有提案，没有关系——所以这里\n产出的仍然是一个要每个人各自点头的提案。\"一键重新邀请\"指的是\n\"一键问他们\"，不是\"一键把他们算进去\"。\n\n仍然过求解器和稳定性检查：上次一起做成过，不等于这次这几个人凑得成\n一个组。跳过这两步等于给他们一个看起来成立、实际不成立的提案。",
        "operationId": "ask_them_again_api_events__event_id__again_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          },
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgainRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Ask Them Again Api Events  Event Id  Again Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/coming-up": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Coming Up",
        "description": "我这两天要出发的事。\n\n只看**已经定下来时间**的（行动确认卡写了 `starts_at`）。没定时间的事\n不该出现在这里——它不是\"快到了\"，它是\"还没定\"，而那一句在别处说。\n\n窗口是从现在到 36 小时后：够覆盖\"明天出发\"，又不会把下周的事提前\n塞进来。提前太多的提醒和不提醒一样，人会学会忽略它。",
        "operationId": "coming_up_api_me_coming_up_get",
        "parameters": [
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          },
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComingUpOut"
                  },
                  "title": "Response Coming Up Api Me Coming Up Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgainRequest": {
        "properties": {
          "intent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Intent Id"
          },
          "invite": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Invite"
          }
        },
        "type": "object",
        "required": [
          "intent_id"
        ],
        "title": "AgainRequest"
      },
...
      "EnvelopeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "intent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Intent Id"
          },
          "grants": {
            "items": {
              "$ref": "#/components/schemas/GrantOut"
            },
            "type": "array",
            "title": "Grants"
          },
          "cited_facet_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Cited Facet Ids"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "consent_record_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Record Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "intent_id",
          "grants",
          "cited_facet_ids",
          "state",
          "expires_at"
        ],
        "title": "EnvelopeOut"
      },
...
      "GrantOut": {
        "properties": {
          "field_name": {
            "type": "string",
            "title": "Field Name"
          },
          "audience": {
            "type": "string",
            "title": "Audience"
          },
          "purposes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Purposes"
          },
          "purpose_label": {
            "type": "string",
            "title": "Purpose Label"
          }
        },
        "type": "object",
        "required": [
          "field_name",
          "audience",
          "purposes",
          "purpose_label"
        ],
        "title": "GrantOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}